+++ source/isaaclab/isaaclab/sim/spawners/wrappers/wrappers.py
@@ -5,17 +5,11 @@
 
 from __future__ import annotations
 
-import random
-import re
 from typing import TYPE_CHECKING
 
-import carb
 import isaacsim.core.utils.prims as prim_utils
-import isaacsim.core.utils.stage as stage_utils
-from isaacsim.core.utils.stage import get_current_stage
-from pxr import Sdf, Usd
+from pxr import Usd
 
-import isaaclab.sim as sim_utils
 from isaaclab.sim.spawners.from_files import UsdFileCfg
 
 if TYPE_CHECKING:
@@ -47,31 +41,6 @@ def spawn_multi_asset(
     Returns:
         The created prim at the first prim path.
     """
-    # get stage handle
-    stage = get_current_stage()
-
-    # resolve: {SPAWN_NS}/AssetName
-    # note: this assumes that the spawn namespace already exists in the stage
-    root_path, asset_path = prim_path.rsplit("/", 1)
-    # check if input is a regex expression
-    # note: a valid prim path can only contain alphanumeric characters, underscores, and forward slashes
-    is_regex_expression = re.match(r"^[a-zA-Z0-9/_]+$", root_path) is None
-
-    # resolve matching prims for source prim path expression
-    if is_regex_expression and root_path != "":
-        source_prim_paths = sim_utils.find_matching_prim_paths(root_path)
-        # if no matching prims are found, raise an error
-        if len(source_prim_paths) == 0:
-            raise RuntimeError(
-                f"Unable to find source prim path: '{root_path}'. Please create the prim before spawning."
-            )
-    else:
-        source_prim_paths = [root_path]
-
-    # find a free prim path to hold all the template prims
-    template_prim_path = stage_utils.get_next_free_path("/World/Template")
-    prim_utils.create_prim(template_prim_path, "Scope")
-
     # spawn everything first in a "Dataset" prim
     proto_prim_paths = list()
     for index, asset_cfg in enumerate(cfg.assets_cfg):
@@ -88,7 +57,7 @@ def spawn_multi_asset(
             if hasattr(asset_cfg, attr_name) and attr_value is not None:
                 setattr(asset_cfg, attr_name, attr_value)
         # spawn single instance
-        proto_prim_path = f"{template_prim_path}/Asset_{index:04d}"
+        proto_prim_path = prim_path.replace(".*", f"{index}")
         asset_cfg.func(
             proto_prim_path,
             asset_cfg,
@@ -100,35 +69,7 @@ def spawn_multi_asset(
         # append to proto prim paths
         proto_prim_paths.append(proto_prim_path)
 
-    # resolve prim paths for spawning and cloning
-    prim_paths = [f"{source_prim_path}/{asset_path}" for source_prim_path in source_prim_paths]
-
-    # manually clone prims if the source prim path is a regex expression
-    # note: unlike in the cloner API from Isaac Sim, we do not "reset" xforms on the copied prims.
-    #   This is because the "spawn" calls during the creation of the proto prims already handles this operation.
-    with Sdf.ChangeBlock():
-        for index, prim_path in enumerate(prim_paths):
-            # spawn single instance
-            env_spec = Sdf.CreatePrimInLayer(stage.GetRootLayer(), prim_path)
-            # randomly select an asset configuration
-            if cfg.random_choice:
-                proto_path = random.choice(proto_prim_paths)
-            else:
-                proto_path = proto_prim_paths[index % len(proto_prim_paths)]
-            # copy the proto prim
-            Sdf.CopySpec(env_spec.layer, Sdf.Path(proto_path), env_spec.layer, Sdf.Path(prim_path))
-
-    # delete the dataset prim after spawning
-    prim_utils.delete_prim(template_prim_path)
-
-    # set carb setting to indicate Isaac Lab's environments that different prims have been spawned
-    # at varying prim paths. In this case, PhysX parser shouldn't optimize the stage parsing.
-    # the flag is mainly used to inform the user that they should disable `InteractiveScene.replicate_physics`
-    carb_settings_iface = carb.settings.get_settings()
-    carb_settings_iface.set_bool("/isaaclab/spawn/multi_assets", True)
-
-    # return the prim
-    return prim_utils.get_prim_at_path(prim_paths[0])
+    return prim_utils.get_prim_at_path(proto_prim_paths[0])
 
 
 def spawn_multi_usd_file(
@@ -178,8 +119,6 @@ def spawn_multi_usd_file(
     for usd_path in usd_paths:
         usd_cfg = usd_template_cfg.replace(usd_path=usd_path)
         multi_asset_cfg.assets_cfg.append(usd_cfg)
-    # set random choice
-    multi_asset_cfg.random_choice = cfg.random_choice
 
     # propagate the contact sensor settings
     # note: the default value for activate_contact_sensors in MultiAssetSpawnerCfg is False.