+++ source/isaaclab/test/sim/test_stage_in_memory.py
@@ -54,16 +54,12 @@ def test_stage_in_memory_with_shapes(sim):
     if isaac_sim_version < 5:
         pytest.skip("Stage in memory is not supported in this version of Isaac Sim")
 
-    # define parameters
-    num_clones = 10
-
     # grab stage in memory and set as current stage via the with statement
     stage_in_memory = sim.get_initial_stage()
     with sim_utils.use_stage(stage_in_memory):
         # create cloned cone stage
-        for i in range(num_clones):
-            prim_utils.create_prim(f"/World/env_{i}", "Xform", translation=(i, i, 0))
-
+        prim_utils.create_prim("/World/Cone", "Xform")
+        num_shape_prototypes = 3
         cfg = sim_utils.MultiAssetSpawnerCfg(
             assets_cfg=[
                 sim_utils.ConeCfg(
@@ -84,7 +80,7 @@ def test_stage_in_memory_with_shapes(sim):
             mass_props=sim_utils.MassPropertiesCfg(mass=1.0),
             collision_props=sim_utils.CollisionPropertiesCfg(),
         )
-        prim_path_regex = "/World/env_.*/Cone"
+        prim_path_regex = "/World/Cone/asset_.*"
         cfg.func(prim_path_regex, cfg)
 
         # verify stage is in memory
@@ -92,13 +88,13 @@ def test_stage_in_memory_with_shapes(sim):
 
         # verify prims exist in stage in memory
         prims = prim_utils.find_matching_prim_paths(prim_path_regex)
-        assert len(prims) == num_clones
+        assert len(prims) == num_shape_prototypes
 
         # verify prims do not exist in context stage
         context_stage = omni.usd.get_context().get_stage()
         with sim_utils.use_stage(context_stage):
             prims = prim_utils.find_matching_prim_paths(prim_path_regex)
-            assert len(prims) != num_clones
+            assert len(prims) != num_shape_prototypes
 
         # attach stage to context
         sim_utils.attach_stage_to_usd_context()
@@ -108,7 +104,7 @@ def test_stage_in_memory_with_shapes(sim):
 
     # verify prims now exist in context stage
     prims = prim_utils.find_matching_prim_paths(prim_path_regex)
-    assert len(prims) == num_clones
+    assert len(prims) == num_shape_prototypes
 
 
 def test_stage_in_memory_with_usds(sim):
@@ -120,7 +116,7 @@ def test_stage_in_memory_with_usds(sim):
         pytest.skip("Stage in memory is not supported in this version of Isaac Sim")
 
     # define parameters
-    num_clones = 10
+    num_robot_prototypes = 2
     usd_paths = [
         f"{ISAACLAB_NUCLEUS_DIR}/Robots/ANYbotics/ANYmal-C/anymal_c.usd",
         f"{ISAACLAB_NUCLEUS_DIR}/Robots/ANYbotics/ANYmal-D/anymal_d.usd",
@@ -129,9 +125,8 @@ def test_stage_in_memory_with_usds(sim):
     # grab stage in memory and set as current stage via the with statement
     stage_in_memory = sim.get_initial_stage()
     with sim_utils.use_stage(stage_in_memory):
-        # create cloned robot stage
-        for i in range(num_clones):
-            prim_utils.create_prim(f"/World/env_{i}", "Xform", translation=(i, i, 0))
+        # create robot stage
+        prim_utils.create_prim("/World/Robot", "Xform")
 
         cfg = sim_utils.MultiUsdFileCfg(
             usd_path=usd_paths,
@@ -150,7 +145,7 @@ def test_stage_in_memory_with_usds(sim):
             ),
             activate_contact_sensors=True,
         )
-        prim_path_regex = "/World/env_.*/Robot"
+        prim_path_regex = "/World/Robot/asset_.*"
         cfg.func(prim_path_regex, cfg)
 
         # verify stage is in memory
@@ -158,13 +153,13 @@ def test_stage_in_memory_with_usds(sim):
 
         # verify prims exist in stage in memory
         prims = prim_utils.find_matching_prim_paths(prim_path_regex)
-        assert len(prims) == num_clones
+        assert len(prims) == num_robot_prototypes
 
         # verify prims do not exist in context stage
         context_stage = omni.usd.get_context().get_stage()
         with sim_utils.use_stage(context_stage):
             prims = prim_utils.find_matching_prim_paths(prim_path_regex)
-            assert len(prims) != num_clones
+            assert len(prims) != num_robot_prototypes
 
         # attach stage to context
         sim_utils.attach_stage_to_usd_context()
@@ -174,7 +169,7 @@ def test_stage_in_memory_with_usds(sim):
 
     # verify prims now exist in context stage
     prims = prim_utils.find_matching_prim_paths(prim_path_regex)
-    assert len(prims) == num_clones
+    assert len(prims) == num_robot_prototypes
 
 
 def test_stage_in_memory_with_clone_in_fabric(sim):
@@ -212,7 +207,6 @@ def test_stage_in_memory_with_clone_in_fabric(sim):
             base_env_path=base_env_path,
             prim_paths=target_paths,
             replicate_physics=True,
-            clone_in_fabric=True,
         )
         prim_path_regex = "/World/envs/env_.*"